requirements.txt
flask==3.0.0
flask-cors==5.0.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
python-dotenv==1.0.0
openpyxl==3.1.2
gunicorn==21.2.0
APScheduler>=3.10.0
pytest==8.3.4
pytest-cov==6.0.0
